changed price to double

+++ GroupProject/SimulationBackup.cpp
@@ -1,6 +1,6 @@
 #include "SimulationBackup.h"
 
-SimulationBackup::SimulationBackup(Rocket* rocket,float price)
+SimulationBackup::SimulationBackup(Rocket* rocket,double price)
 {
 	this->myRocket = rocket->clone();
 	this->price = price;
